+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to the "emacs-mcx" extension will be documented in this file
 ### Fix
 - Drop `C-x C-k`, #1180.
 
+### Add
+- `emacs-mcx.paredit.backwardKillSexp`, #1175.
+
 ## [0.37.1] - 2021-01-01
 ### Fix
 - Assign `C-i` to TAB, #1151.