+++ CHANGELOG.md
@@ -5,8 +5,10 @@ All notable changes to the "emacs-mcx" extension will be documented in this file
 ### Add
 - `emacs-mcx.paredit.backwardKillSexp`, #1175.
 - `emacs-mcx.paredit.markSexp`, #1181.
+
 ### Fix
 - Drop `C-x C-k`, #1180.
+- Cursor positions after `kill-region` in rect-mark-mode, #1187.
 
 ## [0.37.1] - 2021-01-01
 ### Fix